+++ src/vite-end.d.ts
@@ -1,3 +1,7 @@
 /// <reference types="vite/client" />
 declare const GITHUB_RUNTIME_PERMANENT_NAME: string
 declare const BASE_KV_SERVICE_URL: string
+
+interface ImportMetaEnv {
+  readonly VITE_COMMUNITY_STATS_API_ENABLED?: string
+}